+++ srcs/sort/sort_strategy.c
@@ -1,6 +1,6 @@
 #include "push_swap.h"
 
-void    sort_strategy(int disorder, t_stack **stack_a, t_stack **stack_b,
+void    sort_strategy(double disorder, t_stack **stack_a, t_stack **stack_b,
         t_strategy *opt)
 {
     if (ft_strncmp(opt->sort_mode, "adaptative", 11) == 0)
